requirements.txt
ffmpeg-python
openai
discord.py[voice]
sqlmodel
